First, non working yet implementation of serial communication in QtGui. Fixed an integer/pointer conversion in xeq.c Added a missing " in compile_cats.c
git-svn-id: http://svn.code.sf.net/p/wp34s/code@2166 9654ca62-1852-461c-aa49-1be32f3bb02d

+++ trunk/xeq.c
@@ -3131,7 +3131,7 @@ void busy(void)
  */
 static const s_opcode *check_for_xrom_address(void *fp)
 {
-	const s_opcode *xp = (const s_opcode *) ((unsigned long) fp & ~1);
+	const s_opcode *xp = (const s_opcode *) ((uintptr_t ) fp & ~1);
 	if (xp < xrom)
 		return NULL;
 #ifndef REALBUILD
